requirements.txt
# Service Provider Application - Clean Build
# Python 3.14+
# PostgreSQL 18+

# Core
Django>=6.0,<6.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# API
djangorestframework>=3.16.0
djangorestframework-simplejwt>=5.5.0
django-cors-headers>=4.9.0
django-filter>=25.0

# Utilities
requests>=2.31.0
